backend/requirements.txt
aiogram

# sys 
loguru

# .env
python-dotenv